+++ BFSDatabaseGenerator.hh
@@ -4,6 +4,7 @@
 #include <vector>
 #include <map>
 #include <memory>
+#include <cstdint>
 
 #include "Board.hh"
 
@@ -12,25 +13,25 @@ namespace TaquinSolve
     class BFSDatabaseGenerator
     {
         public:
-            void generate(std::vector<size_t> goal_board, std::set<size_t> group_tiles, size_t board_size, std::string output_file);
+            void generate(std::vector<uint8_t> goal_board, std::set<uint8_t> group_tiles, uint8_t board_size, std::string output_file);
 
             std::vector< std::shared_ptr<Board> > perform_moves(Board *board, std::vector<Moves> moves);
 
         protected:
-            std::set<size_t> visited;
-            std::map<size_t, int> database;
+            std::set<uint64_t> visited;
+            std::map<uint64_t, uint8_t> database;
 
             void database_clear();
 
             void database_insert(
                 std::shared_ptr<Board> board,
-                std::shared_ptr< std::set<size_t> > group_tiles_nozero,
-                std::shared_ptr< std::set<size_t> > group_tiles
+                std::shared_ptr< std::set<uint8_t> > group_tiles_nozero,
+                std::shared_ptr< std::set<uint8_t> > group_tiles
             );
 
-            bool check_visited(size_t index);
+            bool check_visited(uint64_t index);
 
-            int database_get_value(size_t index);
+            uint8_t database_get_value(uint64_t index);
 
             void save_database(std::string output_file);
     };
